Modification of variable type for the Rang table

+++ models/ratings.model.js
@@ -16,23 +16,23 @@ module.exports = ( sequelize ) => {
             autoIncrement: true
         },
         rate_picture: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT,
             allowNull: false
         },
         rate_actor_game: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT,
             allowNull: false
         },
         rate_cinematography: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT,
             allowNull: false
         },
         rate_sound: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT,
             allowNull: false
         },
         rate_writing: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.FLOAT,
             allowNull: false
         },
         ID_Movie: {
